+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-### NEXT_VERSION_TYPE=MAJOR
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [1.0.0](https://github.com/yoomoney/db-queue-scheduler/pull/5) (29-10-2021)
+
 * Toggling scheduled tasks feature implemented - `ScheduledTaskSettings.enabled`;
 * Improved library api usage experience - look at the breaking changes;
 * **breaking changes**
 * `ru.yoomoney.tech.dbqueue.scheduler.Scheduler.schedule` method argument order changed;
 * `ru.yoomoney.tech.dbqueue.scheduler.models.ScheduledTask.getIdentity` method added;
 * `scheduledTaskIdentity` field pruned from `ScheduledTaskSettings`.
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [0.3.0](https://github.com/yoomoney/db-queue-scheduler/pull/4) (28-10-2021)
 
 * `ScheduledTaskLifecycleListener` introduced - that helps to observe task execution, eg: tracing, logging, monitoring, etc.